+++ lace-native/csrc/lace_helpers.c
@@ -30,3 +30,17 @@ int lace_is_stolen_task_ext(lace_task *t) {
 int lace_is_completed_task_ext(lace_task *t) {
     return lace_is_completed_task(t);
 }
+
+void *lace_task_result_ext(lace_task *t) {
+    return lace_task_result(t);
+}
+
+void lace_make_all_shared_ext(void) {
+    lace_make_all_shared();
+}
+
+void lace_count_report_ext(void) {
+#ifdef LACE_COUNT_EVENTS
+    lace_count_report();
+#endif
+}